fix: keep timer running across screens
